+++ packages/serverpod_client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.9.8
+- Adds Terraform deployment scripts for AWS. Documentation here: [https://github.com/serverpod/serverpod/discussions/182](https://github.com/serverpod/serverpod/discussions/182)
+- __Breaking change__: Updates structure of config files. Migration instructions here: [https://github.com/serverpod/serverpod/discussions/182](https://github.com/serverpod/serverpod/discussions/182)
+- Moves Redis enabled option to config file and turns it off by default.
+- `serverpod run` no longer manages the Docker containers as it caused an issue with restarting the server.
+
 ## 0.9.7
 - `serverpod create` and `serverpod generate` is now working on Windows. Tested on a fresh install of Windows 10.